+++ src/components/tracker-tabs/Collected.vue
@@ -5,6 +5,8 @@ import { loadCollectedItems, tracker } from '@/state/tracker';
 import check from '@/assets/icons/check.png';
 import PlayerName from '../text-elements/PlayerName.vue';
 import ItemName from '../text-elements/ItemName.vue';
+import { settings } from '@/state/settings';
+import type { ItemClass } from '@/lib/archipelago';
 
 onActivated(() => {
   loadCollectedItems();
@@ -23,15 +25,48 @@ const search = ref('');
 const filteredCollected = computed(() => {
   return tracker.collected.filter(item => item.name.toLowerCase().includes(search.value.toLowerCase()));
 });
+
+const simplifiedColumns: Column[] = [
+  { label: 'Quantity', key: 'count', style: 'width: 80px' },
+  { label: 'Name', key: 'name' }
+];
+
+const simplifiedCollected = computed(() => {
+  const itemCounts: Record<string, { count: number; iclass: ItemClass[]; }> = {};
+
+  for (const item of tracker.collected) {
+    itemCounts[item.name] = itemCounts[item.name] || { count: 0, iclass: item.class };
+    itemCounts[item.name]!.count++;
+  }
+
+  return Object.entries(itemCounts).map(([name, { count, iclass }]) => ({
+    name,
+    count,
+    iclass
+  })).filter(item => item.name.toLowerCase().includes(search.value.toLowerCase()));
+});
 </script>
 
 <template>
   <div class="collected">
     <div>
       <input v-model="search" style="margin-right: auto" placeholder="Search items by name..." type="text">
+      <div class="check-row" style="margin: 0.5em 0;">
+        <input v-model="settings.trackerCollectedSimplified" type="checkbox" id="simplified">
+        <label for="simplified">Simplified view</label>
+      </div>
     </div>
     <div class="sunken-panel">
-      <AppTable :columns="columns" :data="filteredCollected" default-sort-by="order" default-sort-order="desc">
+      <AppTable v-if="settings.trackerCollectedSimplified" :columns="simplifiedColumns" :data="simplifiedCollected" default-sort-by="count" default-sort-order="desc">
+        <template #count="{ item }">
+          <td style="text-align: center">{{ item.count }}</td>
+        </template>
+
+        <template #name="{ item }">
+          <td><ItemName :iclass="item.iclass" :name="item.name" /></td>
+        </template>
+      </AppTable>
+      <AppTable v-else :columns="columns" :data="filteredCollected" default-sort-by="order" default-sort-order="desc">
         <template #_check_image="{ item }">
           <td><img :src="check"></td>
         </template>